% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_on_column.R
\name{comment_on_column}
\alias{comment_on_column}
\title{Add Comment to Database Column}
\usage{
comment_on_column(conn, table, column, comment)
}
\arguments{
\item{conn}{A database connection object (DBI connection)}

\item{table}{Character string specifying the table name}

\item{column}{Character string specifying the column name}

\item{comment}{Character string with the comment text to add}
}
\value{
Number of rows affected (typically 0 for DDL statements)
}
\description{
Adds a comment to a specific column in a database table using SQL COMMENT statement.
}
\examples{
\dontrun{
library(DBI)
conn <- dbConnect(...)
comment_on_column(conn, "my_table", "my_column", "This is a description")
}

}
