% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment_on_table.R
\name{comment_on_table}
\alias{comment_on_table}
\title{Add Comment to Database Table}
\usage{
comment_on_table(conn, table, comment)
}
\arguments{
\item{conn}{A database connection object (DBI connection)}

\item{table}{Character string specifying the table name}

\item{comment}{Character string with the comment text to add}
}
\value{
Number of rows affected (typically 0 for DDL statements)
}
\description{
Adds a comment to a database table using SQL COMMENT statement.
}
\examples{
\dontrun{
library(DBI)
conn <- dbConnect(...)
comment_on_table(conn, "my_table", "This table contains user data")
}

}
